[fix-typescript-error] fixng mocha, nyc, and eslint
