+++ .mocharc.json
@@ -1,11 +1,12 @@
 {
   "diff": true,
-  "extension": ["ts", "tx", "mjs"],
-  "package": "./package.json",
+  "extension": ["ts", "tx"],
   "spec": "./tests/**/*.ts",
   "reporter": "spec",
   "slow": "75",
   "timeout": "2000",
   "ui": "bdd",
-  "recursive": true
+  "recursive": true,
+  "require": "ts-node/register",
+  "exclude": ["dist"]
 }
