+++ index.d.ts
@@ -1,0 +1,9 @@
+export { };
+
+declare global {
+  namespace Express {
+    export interface Request {
+      id: string;
+    }
+  }
+}
