--- src/types/index.d.ts
@@ -1,18 +1,0 @@
-import type { Request } from 'express-serve-static-core';
-
-export { }
-
-declare global {
-  namespace Express {
-    export interface Request {
-      id: string;
-    }
-  }
-}
-
-export type Options = {
-  generator?: ((_request: Request) => string);
-  headerName?: string;
-  setHeader?: boolean;
-  attrName?: string;
-}
